Clean up plans types
